Some refactoring in neural_networks; added script to pretrain CNN and functioanlity to visualize vision_encoder in jupiter notebook. Also added functionality to be able to continue dataset-recording.

+++ trackmania_env/envs/testcases_single_agent.py
@@ -463,7 +463,14 @@ class Test_Reward_Next_Point_Manager(TestEnvironmentCallback):
 
 
 class PretrainingDataCollection(TestEnvironmentCallback):
-    def __init__(self, reference_line_manager : ReferenceLineManager, logging_directory : str):
+    def __init__(self, reference_line_manager : ReferenceLineManager, logging_directory : str, continuation_idx : int = -1):
+        """
+        Parameters
+        -----------
+            - reference_line_manager    : ReferenceLineManager used by environment
+            - logging_directory         : Logging directory in which dataset is created or recorded
+            - continueation_idx         : If this is set to some index other than -1, it is assumed there is already an existing dataset and
+                                            the existing dataset is upposed to be extended. If it's -1, a new dataset is created (be careful; if it's -1 and there IS a dataset there, it'll be overwritten.)"""
         super().__init__()
         self.ref_line_manager = reference_line_manager
         self.logging_directory = logging_directory
@@ -471,9 +478,14 @@ class PretrainingDataCollection(TestEnvironmentCallback):
         os.makedirs(self.img_directory, exist_ok = True)
         self.labels = os.path.join(self.logging_directory, "labels.csv")
         
-        if not os.path.exists(self.labels):
+        if not os.path.exists(self.labels) and continuation_idx == -1:
             with open(self.labels, "w") as file:
                 file.write("filename,lateral_distance\n")
+
+        if not continuation_idx == -1:
+            self.n_step = continuation_idx
+        else:
+            assert os.path.exists(self.labels), "Expected dataset to already exists but did not find labels."
 
 
     def _call_after_step(self, processed_obs, reward, terminated, truncated, info):
@@ -491,6 +503,5 @@ class PretrainingDataCollection(TestEnvironmentCallback):
         img : np.ndarray = processed_obs["image"]
         np.save(os.path.join(self.img_directory, f"img_{self.n_step}.npy"), img)
 
-        
         
         self.n_step += 1
